@@ -22,11 +22,10 @@
 		"i18next-multiload-backend-adapter": "^2.3.0",
 		"lucide-react": "^0.563.0",
 		"qrcode.react": "^1.0.1",
-		"react": "^16.14.0",
+		"react": "^18.3.1",
 		"react-chartjs-2": "^4.2.0",
-		"react-dom": "npm:@hot-loader/react-dom",
+		"react-dom": "^18.3.1",
 		"react-fast-compare": "^3.2.0",
-		"react-hot-loader": "^4.12.21",
 		"react-hot-toast": "^2.6.0",
 		"react-i18next": "^16.5.4",
 		"react-router-dom": "^5.1.2",
@@ -66,8 +65,8 @@
 		"@types/debounce": "^1.2.0",
 		"@types/node": "^25.2.2",
 		"@types/qrcode.react": "^1.0.1",
-		"@types/react": "^16.14.0",
-		"@types/react-dom": "^16.9.16",
+		"@types/react": "^18.3.1",
+		"@types/react-dom": "^18.3.1",
 		"@types/react-router": "^5.1.3",
 		"@types/react-router-dom": "^5.1.3",
 		"@types/react-transition-group": "^4.4.0",
