@@ -1,4 +1,5 @@
 import type React from "react";
+import type { PropsWithChildren } from "react";
 import Can from "@/components/elements/Can";
 import { ServerError } from "@/components/elements/ScreenBlock";
 
@@ -6,10 +7,9 @@ export interface RequireServerPermissionProps {
 	permissions: string | string[];
 }
 
-const RequireServerPermission: React.FC<RequireServerPermissionProps> = ({
-	children,
-	permissions,
-}) => {
+const RequireServerPermission: React.FC<
+	PropsWithChildren<RequireServerPermissionProps>
+> = ({ children, permissions }) => {
 	return (
 		<Can
 			action={permissions}
